+++ src/state/slices/opportunitiesSlice/resolvers/thorchainsavers/types.ts
@@ -9,3 +9,33 @@ export type ThorchainSaversStakingSpecificMetadata = OpportunityMetadataBase & {
   saversMaxSupplyFiat: string
   isFull: boolean
 }
+
+export type MidgardPoolResponse = {
+  annualPercentageRate: string
+  asset: string
+  assetDepth: string
+  assetPrice: string
+  assetPriceUSD: string
+  liquidityUnits: string
+  nativeDecimal: string
+  poolAPY: string
+  runeDepth: string
+  saversAPR: string
+  saversDepth: string
+  saversUnits: string
+  status: string
+  synthSupply: string
+  synthUnits: string
+  units: string
+  volume24h: string
+}
+
+export type ThorchainSaverPositionResponse = {
+  asset: string
+  asset_address: string
+  last_add_height: number
+  units: string
+  asset_deposit_value: string
+  asset_redeem_value: string
+  growth_pct: string
+}
